vaibhavanidhi-backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
bcrypt==4.0.1
python-dotenv==1.0.0
datetime==5.1
Flask-Migrate==4.0.5